requirements.txt
Django>=1.8,<1.9
djangorestframework>=3.1,<4.0
six
